optimised code with proper naming conventions and absence of unnecessary folders, files and code
