+++ src/Api/createProducts.js
@@ -12,8 +12,8 @@ export const CreateProduct = async (id, image, name, description, unit, category
                     region: 'us-east-1',
                     defaultAuthMode: 'apiKey',
                     apiKey: 'da2-tt7a24loa5ch7ceq7onemeej7a'
-                }
-              
+             }
+
             }
         });
