+++ src/Api/fetchingOrders.js
@@ -18,8 +18,7 @@ const client = generateClient();
                 region: 'us-east-1',
                 defaultAuthMode: 'apiKey',
                 apiKey: 'da2-tt7a24loa5ch7ceq7onemeej7a'
-            }
-          
+        }
 
           }
         });
