+++ src/Api/fetchingcustomers.js
@@ -18,8 +18,7 @@ const client = generateClient();
                 region: 'us-east-1',
                 defaultAuthMode: 'apiKey',
                 apiKey: 'da2-tt7a24loa5ch7ceq7onemeej7a'
-            }
-          
+         }
 
           }
         });
@@ -50,11 +49,3 @@ const client = generateClient();
           console.log(error)
       }
     }
-//   fetchCategories()
-
-//   const [Data, setData] = useState([])
-//   console.log(Data);
-
-//   const uniqueCategories = [...new Set(Data.map((item) => item.category))];
-
-//   console.log(uniqueCategories);
