+++ src/app/admin/productlist/page.js
@@ -1,5 +1,5 @@
 'use client'
-import ProductList from '@/components/admin/productsList/Pro'
+import ProductList from '@/components/admin/print/print'
 import React from 'react'
 
 const page = () => {
