+++ src/components/admin/print/print.js
