+++ src/components/admin/share/Share.js
@@ -181,12 +181,6 @@ const Share = () => {
       )
     }
   ];
-    const rowSelection = {
-        onChange: (selectedRowKeys, selectedRows) => {
-          console.log("selectedRowKeys:", selectedRowKeys);
-          console.log("selectedRows:", selectedRows);
-        },
-      };
       const currentDate = new Date().toLocaleDateString();
       const currentTime = new Date().toLocaleTimeString([], {
         hour: '2-digit',
@@ -296,7 +290,7 @@ const Share = () => {
 
 
       return (
-        <>
+        <div>
           <button
             key="link"
           
@@ -326,7 +320,7 @@ const Share = () => {
             Successfully Shared
           </p>
      </div>)}
-  </>
+  </div>
 );
 };
 export default Share;
